refactor(custom_tools): rename _google_oauth.py to google_oauth.py
Three sibling custom_tools modules (gmail, google_calendar, google_sheets)
import this file as their OAuth helper. Per the naming convention "private
iff nothing under src/ imports it", it is de-facto public and should drop
the leading underscore. Adding the tach.toml [[modules]] declaration is
deferred to ARCH-B-8 cross-cutting cleanup.
